+++ policyengine-client/src/policyengine/pages/populationImpact/populationImpact.jsx
@@ -11,6 +11,7 @@ import Spinner from "../../general/spinner";
 import { useContext } from "react";
 import NavigationButton from "../../general/navigationButton";
 import { OverviewHolder, PolicyOverview, SharePolicyLinks } from "../policy/overview";
+import AgeChart from "./ageChart";
 function PopulationResultsCaveats() {
 	return <p style={{ color: "grey" }}><ExclamationCircleOutlined />  &nbsp; &nbsp;PolicyEngine results assume no behavioural or macroeconomic effects</p>;
 }
@@ -135,6 +136,9 @@ export function PopulationResultsPane(props) {
 			</Row>
 			<Row>
 				<BreakdownTable policy={country.policy} api_url={country.apiURL} />
+			</Row>
+			<Row>
+				<AgeChart policy={country.policy} api_url={country.apiURL} />
 			</Row>
 			<Divider />
 			<PopulationResultsCaveats />